refactor: replace directory path
디렉터리 구조 변경으로 인한 수정
